Refactor : Added new Difficulty Model and also structured the compoents
