+++ src/app/Menu.tsx
@@ -6,7 +6,7 @@ import { useCoins, useXP, useUser, useMute, useTut } from '@/services/store';
 import TutorialModal from '../modals/TutorialModal';
 import { toast } from "react-toastify";
 import { useToastCooldown } from "@/components/hooks/useToastCooldown";
-import { MenuButton } from '@/components/ui/MenuButton';
+import { MenuButton } from '@/components/ui/Buttons/MenuButton';
 
 const Menu = () => {
   const setCoins = useCoins((state) => state.setCoins);
