+++ src/modals/DifficultyModal.tsx
@@ -1,4 +1,5 @@
 import { DifficultyLevel, DifficultyModalProps } from "../services/types";
+import { DifficultyActionButton } from "@/components/ui/Buttons/DifficultyActionButton";
 
 const DifficultyModal = ({ visible, onSelect, onClose }: DifficultyModalProps) => {
   if (!visible) return null;
@@ -9,21 +10,21 @@ const DifficultyModal = ({ visible, onSelect, onClose }: DifficultyModalProps) =
         <h2 className="text-white text-4xl text-center mb-6">Select Difficulty</h2>
 
         {[1, 2, 3, 4, 5].map(level => (
-          <button
+          <DifficultyActionButton
+            variant="level"
             key={level}
             onClick={() => onSelect(level as DifficultyLevel)}
-            className="w-full py-3 my-2 bg-blue-600 text-white text-3xl hover:bg-blue-700 transition"
           >
             Level {level}
-          </button>
+          </DifficultyActionButton>
         ))}
 
-        <button
+        <DifficultyActionButton
+          variant="cancel"
           onClick={onClose}
-          className="w-full mt-4 py-3 bg-red-600 text-white text-3xl hover:bg-red-700 transition"
         >
           Cancel
-        </button>
+        </DifficultyActionButton>
       </div>
     </div>
   );
